Align with DROOLS-1434 for moving public interfaces to kie-dmn-api module
