+++ kie-server-parent/kie-server-api/pom.xml
@@ -50,11 +50,7 @@
     <!-- Needed only for the DMNResult, DMNContext, etc and FEELEvent so should be possible later to migrate to kie-dmn-API -->
     <dependency>
       <groupId>org.kie</groupId>
-      <artifactId>kie-dmn-core</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.kie</groupId>
-      <artifactId>kie-dmn-feel</artifactId>
+      <artifactId>kie-dmn-api</artifactId>
     </dependency>
 
     <!-- json -->
