+++ kie-server-parent/kie-server-api/src/main/java/org/kie/server/api/model/dmn/DMNDecisionResultKS.java
@@ -9,8 +9,8 @@ import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlElementWrapper;
 import javax.xml.bind.annotation.XmlRootElement;
 
-import org.kie.dmn.core.api.DMNDecisionResult;
-import org.kie.dmn.core.api.DMNMessage;
+import org.kie.dmn.api.core.DMNDecisionResult;
+import org.kie.dmn.api.core.DMNMessage;
 
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlRootElement(name = "dmn-decision-result")
