+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-dmn/src/test/java/org/kie/server/integrationtests/dmn/DMNIntegrationTest.java
@@ -18,9 +18,9 @@ package org.kie.server.integrationtests.dmn;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.kie.api.KieServices;
-import org.kie.dmn.core.api.DMNContext;
+import org.kie.dmn.api.core.DMNContext;
 import org.kie.dmn.core.api.DMNFactory;
-import org.kie.dmn.core.api.DMNResult;
+import org.kie.dmn.api.core.DMNResult;
 import org.kie.server.api.model.ReleaseId;
 import org.kie.server.api.model.ServiceResponse;
 import org.kie.server.api.model.dmn.DMNResultKS;
